Move js functions to new file.
